+++ src/lib.rs
@@ -33,7 +33,7 @@
 //!     storage.init().await.unwrap();
 //!     let key = ArrayKey::<8>::default();
 //!     let data = b"Hello World!".to_vec();
-//!     storage.write(key, data).await.unwrap();
+//!     storage.write(key, data.into()).await.unwrap();
 //! }
 //! ```
 
@@ -44,6 +44,7 @@ extern crate serde_derive;
 #[macro_use]
 extern crate anyhow;
 
+extern crate bytes;
 extern crate ring;
 
 /// Basic info about current build.