+++ src/storage/core.rs
@@ -4,6 +4,7 @@ use crate::{
 };
 
 use super::prelude::*;
+use bytes::Bytes;
 use futures::stream::FuturesOrdered;
 use tokio::fs::{create_dir, create_dir_all};
 
@@ -53,6 +54,7 @@ where
     pub(crate) safe: Arc<RwLock<Safe<K>>>,
     next_blob_id: Arc<AtomicUsize>,
     pub(crate) ioring: Option<Rio>,
+    pub(crate) corrupted_blobs: Arc<AtomicUsize>,
 }
 
 #[derive(Debug)]
@@ -240,7 +242,7 @@ where
     ///
     /// async fn write_data(storage: Storage<ArrayKey<8>>) {
     ///     let key = ArrayKey::<8>::default();
-    ///     let data = b"async written to blob".to_vec();
+    ///     let data = b"async written to blob".to_vec().into();
     ///     storage.write(key, data).await;
     /// }
     /// ```
@@ -248,7 +250,7 @@ where
     /// Fails with the same errors as [`write_with`]
     ///
     /// [`write_with`]: Storage::write_with
-    pub async fn write(&self, key: impl AsRef<K>, value: Vec<u8>) -> Result<()> {
+    pub async fn write(&self, key: impl AsRef<K>, value: Bytes) -> Result<()> {
         self.write_with_optional_meta(key, value, None).await
     }
 
@@ -259,7 +261,7 @@ where
     ///
     /// async fn write_data(storage: Storage<ArrayKey<8>>) {
     ///     let key = ArrayKey::<8>::default();
-    ///     let data = b"async written to blob".to_vec();
+    ///     let data = b"async written to blob".to_vec().into();
     ///     let mut meta = Meta::new();
     ///     meta.insert("version".to_string(), b"1.0".to_vec());
     ///     storage.write_with(&key, data, meta).await;
@@ -267,7 +269,7 @@ where
     /// ```
     /// # Errors
     /// Fails if duplicates are not allowed and record already exists.
-    pub async fn write_with(&self, key: impl AsRef<K>, value: Vec<u8>, meta: Meta) -> Result<()> {
+    pub async fn write_with(&self, key: impl AsRef<K>, value: Bytes, meta: Meta) -> Result<()> {
         self.write_with_optional_meta(key, value, Some(meta)).await
     }
 
@@ -295,7 +297,7 @@ where
     async fn write_with_optional_meta(
         &self,
         key: impl AsRef<K>,
-        value: Vec<u8>,
+        value: Bytes,
         meta: Option<Meta>,
     ) -> Result<()> {
         let key = key.as_ref();
@@ -319,7 +321,7 @@ where
             .active_blob
             .as_ref()
             .ok_or_else(Error::active_blob_not_set)?;
-        let result = Blob::write(blob, record).await.or_else(|err| {
+        let result = Blob::write(blob, key, record).await.or_else(|err| {
             let e = err.downcast::<Error>()?;
             if let ErrorKind::FileUnavailable(kind) = e.kind() {
                 let work_dir = self
@@ -380,7 +382,7 @@ where
     /// [`Error::RecordNotFound`]: enum.Error.html#RecordNotFound
     /// [`read_with`]: Storage::read_with
     #[inline]
-    pub async fn read(&self, key: impl AsRef<K>) -> Result<Vec<u8>> {
+    pub async fn read(&self, key: impl AsRef<K>) -> Result<Bytes> {
         let key = key.as_ref();
         debug!("storage read {:?}", key);
         self.read_with_optional_meta(key, None).await
@@ -402,7 +404,7 @@ where
     ///
     /// [`Error::RecordNotFound`]: enum.Error.html#RecordNotFound
     #[inline]
-    pub async fn read_with(&self, key: impl AsRef<K>, meta: &Meta) -> Result<Vec<u8>> {
+    pub async fn read_with(&self, key: impl AsRef<K>, meta: &Meta) -> Result<Bytes> {
         let key = key.as_ref();
         debug!("storage read with {:?}", key);
         self.read_with_optional_meta(key, Some(meta))
@@ -445,7 +447,7 @@ where
         Ok(all_entries)
     }
 
-    async fn read_with_optional_meta(&self, key: &K, meta: Option<&Meta>) -> Result<Vec<u8>> {
+    async fn read_with_optional_meta(&self, key: &K, meta: Option<&Meta>) -> Result<Bytes> {
         debug!("storage read with optional meta {:?}, {:?}", key, meta);
         let safe = self.inner.safe.read().await;
         if let Some(ablob) = safe.active_blob.as_ref() {
@@ -460,7 +462,7 @@ where
         Self::get_any_data(&safe, key, meta).await
     }
 
-    async fn get_data_last(safe: &Safe<K>, key: &K, meta: Option<&Meta>) -> Result<Vec<u8>> {
+    async fn get_data_last(safe: &Safe<K>, key: &K, meta: Option<&Meta>) -> Result<Bytes> {
         let blobs = safe.blobs.read().await;
         let possible_blobs = blobs
             .iter_possible_childs_rev(key)
@@ -491,7 +493,7 @@ where
     }
 
     #[allow(dead_code)]
-    async fn get_data_any(safe: &Safe<K>, key: &K, meta: Option<&Meta>) -> Result<Vec<u8>> {
+    async fn get_data_any(safe: &Safe<K>, key: &K, meta: Option<&Meta>) -> Result<Bytes> {
         let blobs = safe.blobs.read().await;
         let stream: FuturesUnordered<_> = blobs
             .iter_possible_childs_rev(key)
@@ -505,7 +507,7 @@ where
             .with_context(|| "no results in closed blobs")
     }
 
-    async fn get_any_data(safe: &Safe<K>, key: &K, meta: Option<&Meta>) -> Result<Vec<u8>> {
+    async fn get_any_data(safe: &Safe<K>, key: &K, meta: Option<&Meta>) -> Result<Bytes> {
         Self::get_data_last(safe, key, meta).await
     }
 
@@ -546,6 +548,11 @@ where
         } else {
             count
         }
+    }
+
+    /// `blob_count` returns exact number of corrupted blobs.
+    pub fn corrupted_blobs_count(&self) -> usize {
+        self.inner.corrupted_blobs.load(Ordering::Acquire)
     }
 
     /// `active_index_memory` returns the amount of memory used by blob to store active indices
@@ -604,6 +611,9 @@ where
     }
 
     async fn init_new(&mut self) -> Result<()> {
+        let corrupted = Self::count_old_corrupted_blobs(&self.inner.config).await;
+        self.inner.corrupted_blobs.store(corrupted, Ordering::Release);
+
         let next = self.inner.next_blob_name()?;
         let mut safe = self.inner.safe.write().await;
         let blob =
@@ -615,7 +625,7 @@ where
     async fn init_from_existing(&mut self, files: Vec<DirEntry>, with_active: bool) -> Result<()> {
         trace!("init from existing: {:#?}", files);
         let disk_access_sem = self.observer.get_dump_sem();
-        let mut blobs = Self::read_blobs(
+        let (mut blobs, corrupted_count) = Self::read_blobs(
             &files,
             self.inner.ioring.clone(),
             disk_access_sem,
@@ -623,6 +633,8 @@ where
         )
         .await
         .context("failed to read blobs")?;
+
+        self.inner.corrupted_blobs.store(corrupted_count, Ordering::Release);
 
         debug!("{} blobs successfully created", blobs.len());
         blobs.sort_by_key(Blob::id);
@@ -669,7 +681,9 @@ where
         ioring: Option<Rio>,
         disk_access_sem: Arc<Semaphore>,
         config: &Config,
-    ) -> Result<Vec<Blob<K>>> {
+    ) -> Result<(Vec<Blob<K>>, usize)> {
+        let mut corrupted = Self::count_old_corrupted_blobs(config).await;
+
         debug!("read working directory content");
         let dir_content = files.iter().map(DirEntry::path);
         debug!("read {} entities", dir_content.len());
@@ -682,6 +696,7 @@ where
                 None
             }
         });
+
         debug!("init blobs from found files");
         let mut futures: FuturesUnordered<_> = blob_files
             .map(|file| async {
@@ -712,13 +727,43 @@ where
                             .with_context(|| {
                                 anyhow!(format!("failed to save corrupted blob {:?}", file))
                             })?;
+                        corrupted += 1;
                     } else {
                         return Err(e.context(msg));
                     }
                 }
             }
         }
-        Ok(blobs)
+        Ok((blobs, corrupted))
+    }
+
+    async fn count_old_corrupted_blobs(config: &Config) -> usize {
+        let mut corrupted = 0;
+
+        if let Some(work_dir_path) = config.work_dir() {
+            let mut corrupted_dir_path = work_dir_path.to_path_buf();
+            corrupted_dir_path.push(config.corrupted_dir_name());
+            if corrupted_dir_path.exists() {
+                let dir = read_dir(&corrupted_dir_path).await;
+                if let Err(e) = dir {
+                    warn!("can't read corrupted blob dir {}: {}", corrupted_dir_path.display(), e);
+                    return corrupted;
+                }
+                let mut dir = dir.unwrap();
+
+                while let Ok(Some(file)) = dir.next_entry().await {
+                    let path = file.path();
+                    if path.is_file() {
+                        let extension = path.extension();
+                        if let Some(BLOB_FILE_EXTENSION) = extension.and_then(|ext| ext.to_str()) {
+                            corrupted += 1;
+                        }
+                    }
+                }
+            }
+        }
+        
+        corrupted
     }
 
     fn should_save_corrupted_blob(error: &anyhow::Error) -> bool {
@@ -904,18 +949,24 @@ where
             .map(|b| b.mark_all_as_deleted(key))
             .collect::<FuturesUnordered<_>>();
         let total = entries_closed_blobs
-            .filter_map(|result| match result {
-                Ok(count) => count,
+            .map(|result| match result {
+                Ok(flag) => {
+                    if flag {
+                        1
+                    } else {
+                        0
+                    }
+                }
                 Err(error) => {
                     warn!("failed to delete records: {}", error);
-                    None
+                    0
                 }
             })
-            .fold(0, |a, b| a + b)
+            .fold(0, |s, n| s + n)
             .await;
         debug!("{} deleted from closed blobs", total);
         self.observer.defer_dump_old_blob_indexes().await;
-        Ok(total)
+        Ok(total as u64)
     }
 
     async fn mark_all_as_deleted_active(&self, key: &K) -> Result<u64> {
@@ -923,7 +974,8 @@ where
         let active_blob = safe.active_blob.as_deref_mut();
         let count = if let Some(active_blob) = active_blob {
             let mut active_blob = active_blob.write().await;
-            let count = active_blob.mark_all_as_deleted(key).await?.unwrap_or(0);
+            let is_deleted = active_blob.mark_all_as_deleted(key).await?;
+            let count = if is_deleted { 1 } else { 0 };
             debug!("{} deleted from active blob", count);
             count
         } else {
@@ -943,6 +995,7 @@ where
             config,
             next_blob_id: Arc::new(AtomicUsize::new(0)),
             ioring,
+            corrupted_blobs: Arc::new(AtomicUsize::new(0)),
         }
     }